requirement.txt
fastapi
uvicorn[standard]
sqlmodel
alembic
psycopg2-binary
python-dotenv
dbfread
pandas
selenium
requests
tqdm
pyproj
